+++ test/migration_v33_test.dart
@@ -85,7 +85,7 @@ void main() {
     expect((await db.getLaciMejaTakenQty(['b-lama']))['b-lama'], 3);
 
     final ver = await db.customSelect('PRAGMA user_version').getSingle();
-    expect(ver.data.values.first, 42);
+    expect(ver.data.values.first, 43);
 
     await db.close();
     if (file.existsSync()) file.deleteSync();